Fix ChargePriceBusinessValidationRulesFactoryTests
