+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Domain/Dtos/ChargeCommands/Validation/BusinessValidation/ChargePriceBusinessValidationRulesFactoryTests.cs
@@ -15,7 +15,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using AutoFixture.Xunit2;
 using FluentAssertions;
@@ -27,7 +26,6 @@ using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands.Validation.BusinessValid
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands.Validation.BusinessValidation.ValidationRules;
 using GreenEnergyHub.Charges.Domain.Dtos.Validation;
 using GreenEnergyHub.Charges.Infrastructure.Core.Cim.ValidationErrors;
-using GreenEnergyHub.Charges.MessageHub.Models.Shared;
 using GreenEnergyHub.Charges.TestCore.Attributes;
 using GreenEnergyHub.Charges.Tests.Builders.Command;
 using Moq;
@@ -36,90 +34,29 @@ using Xunit.Categories;
 
 namespace GreenEnergyHub.Charges.Tests.Domain.Dtos.ChargeCommands.Validation.BusinessValidation
 {
-    // TODO: Adapt test methods to fit with ChargePricesDto
     [UnitTest]
     public class ChargePriceBusinessValidationRulesFactoryTests
     {
         [Theory]
         [InlineAutoMoqData(typeof(StartDateValidationRule))]
-        public async Task CreateRulesAsync_WhenCalledWithNewCharge_ReturnsExpectedMandatoryRules(
+        public async Task CreateRulesAsync_WhenCalledWithNewPrices_ReturnsExpectedMandatoryRules(
             Type expectedRule,
-            [Frozen] Mock<IChargeIdentifierFactory> chargeIdentifierFactory,
-            [Frozen] Mock<IChargeRepository> chargeRepository,
             [Frozen] Mock<IRulesConfigurationRepository> rulesConfigurationRepository,
             ChargePriceBusinessValidationRulesFactory sut,
             ChargePriceDtoBuilder builder)
         {
             // Arrange
             var operation = builder.Build();
-            Charge? charge = null;
 
             SetupConfigureRepositoryMock(rulesConfigurationRepository);
-            SetupChargeIdentifierFactoryMock(chargeIdentifierFactory);
-            SetupChargeRepositoryMock(chargeRepository, charge!);
 
             // Act
             var actual = await sut.CreateRulesAsync(operation).ConfigureAwait(false);
             var actualRules = actual.GetRules().Select(r => r.ValidationRule.GetType());
 
             // Assert
-            actual.GetRules().Count.Should().Be(1); // This assert is added to ensure that when the rule set is expanded, the test gets attention as well.
+            actual.GetRules().Should().HaveCount(1); // This assert is added to ensure that when the rule set is expanded, the test gets attention as well.
             actualRules.Should().Contain(expectedRule);
-        }
-
-        [Theory]
-        [InlineAutoMoqData(typeof(StartDateValidationRule))]
-        [InlineAutoMoqData(typeof(UpdateChargeMustHaveEffectiveDateBeforeOrOnStopDateRule))]
-        [InlineAutoMoqData(typeof(ChargeResolutionCanNotBeUpdatedRule))]
-        public async Task CreateRulesAsync_WhenCalledWithExistingChargeNotTariff_ReturnsExpectedRules(
-            Type expectedRule,
-            [Frozen] Mock<IChargeIdentifierFactory> chargeIdentifierFactory,
-            [Frozen] Mock<IChargeRepository> chargeRepository,
-            [Frozen] Mock<IRulesConfigurationRepository> rulesConfigurationRepository,
-            ChargePriceBusinessValidationRulesFactory sut,
-            Charge charge)
-        {
-            // Arrange
-            var chargeOperationDto = new ChargePriceDtoBuilder().WithChargeType(ChargeType.Fee).Build();
-            SetupConfigureRepositoryMock(rulesConfigurationRepository);
-            SetupChargeIdentifierFactoryMock(chargeIdentifierFactory);
-            SetupChargeRepositoryMock(chargeRepository, charge);
-
-            // Act
-            var actual = await sut.CreateRulesAsync(chargeOperationDto).ConfigureAwait(false);
-            var actualRules = actual.GetRules().Select(r => r.ValidationRule.GetType());
-
-            // Assert
-            Assert.Equal(3, actual.GetRules().Count); // This assert is added to ensure that when the rule set is expanded, the test gets attention as well.
-            Assert.Contains(expectedRule, actualRules);
-        }
-
-        [Theory]
-        [InlineAutoMoqData(typeof(StartDateValidationRule))]
-        [InlineAutoMoqData(typeof(ChangingTariffTaxValueNotAllowedRule))]
-        [InlineAutoMoqData(typeof(UpdateChargeMustHaveEffectiveDateBeforeOrOnStopDateRule))]
-        [InlineAutoMoqData(typeof(ChargeResolutionCanNotBeUpdatedRule))]
-        public async Task CreateRulesAsync_WhenCalledWithExistingTariff_ReturnsExpectedRules(
-            Type expectedRule,
-            [Frozen] Mock<IChargeIdentifierFactory> chargeIdentifierFactory,
-            [Frozen] Mock<IChargeRepository> chargeRepository,
-            [Frozen] Mock<IRulesConfigurationRepository> rulesConfigurationRepository,
-            ChargePriceBusinessValidationRulesFactory sut,
-            Charge charge)
-        {
-            // Arrange
-            var chargeOperationDto = new ChargePriceDtoBuilder().WithChargeType(ChargeType.Tariff).Build();
-            SetupConfigureRepositoryMock(rulesConfigurationRepository);
-            SetupChargeRepositoryMock(chargeRepository, charge);
-            SetupChargeIdentifierFactoryMock(chargeIdentifierFactory);
-
-            // Act
-            var actual = await sut.CreateRulesAsync(chargeOperationDto).ConfigureAwait(false);
-
-            // Assert
-            var actualRules = actual.GetRules().Select(r => r.ValidationRule.GetType());
-            Assert.Equal(4, actual.GetRules().Count); // This assert is added to ensure that when the rule set is expanded, the test gets attention as well.
-            Assert.Contains(expectedRule, actualRules);
         }
 
         [Theory]
@@ -141,18 +78,13 @@ namespace GreenEnergyHub.Charges.Tests.Domain.Dtos.ChargeCommands.Validation.Bus
         [InlineAutoMoqData(CimValidationErrorTextToken.ChargePointPrice)]
         public async Task CreateRulesAsync_WithChargeCommandAllRulesThatNeedTriggeredByForErrorMessage_MustImplementIValidationRuleWithExtendedData(
             CimValidationErrorTextToken cimValidationErrorTextToken,
-            [Frozen] Mock<IChargeIdentifierFactory> chargeIdentifierFactory,
-            [Frozen] Mock<IChargeRepository> chargeRepository,
             [Frozen] Mock<IRulesConfigurationRepository> rulesConfigurationRepository,
             ChargePriceBusinessValidationRulesFactory sut,
             ChargeCommandBuilder chargeCommandBuilder,
-            ChargePriceDtoBuilder chargePriceDtoBuilder,
-            Charge charge)
+            ChargePriceDtoBuilder chargePriceDtoBuilder)
         {
             // Arrange
             SetupConfigureRepositoryMock(rulesConfigurationRepository);
-            SetupChargeIdentifierFactoryMock(chargeIdentifierFactory);
-            SetupChargeRepositoryMock(chargeRepository, charge);
             var chargePriceDto = chargePriceDtoBuilder.Build();
             var chargeCommand = chargeCommandBuilder.WithChargeOperation(chargePriceDto).Build();
 
@@ -164,31 +96,8 @@ namespace GreenEnergyHub.Charges.Tests.Domain.Dtos.ChargeCommands.Validation.Bus
             }
 
             // Assert
-            AssertAllRulesThatNeedTriggeredByForErrorMessageImplementsIValidationRuleWithExtendedData(
+            CimValidationErrorMessageAssertionHelper.AssertAllRulesThatNeedTriggeredByForErrorMessageImplementsIValidationRuleWithExtendedData(
                 cimValidationErrorTextToken, validationRules);
-        }
-
-        private static void AssertAllRulesThatNeedTriggeredByForErrorMessageImplementsIValidationRuleWithExtendedData(
-            CimValidationErrorTextToken cimValidationErrorTextToken,
-            IReadOnlyCollection<IValidationRuleContainer> validationRules)
-        {
-            var type = typeof(CimValidationErrorTextTemplateMessages);
-            foreach (var fieldInfo in type.GetFields(BindingFlags.Static | BindingFlags.Public))
-            {
-                if (!fieldInfo.GetCustomAttributes().Any()) continue;
-
-                var errorMessageForAttribute = (ErrorMessageForAttribute)fieldInfo.GetCustomAttributes()
-                    .Single(x => x.GetType() == typeof(ErrorMessageForAttribute));
-
-                var validationRuleIdentifier = errorMessageForAttribute.ValidationRuleIdentifier;
-                var errorText = fieldInfo.GetValue(null)!.ToString();
-                var validationErrorTextTokens = CimValidationErrorTextTokenMatcher.GetTokens(errorText!);
-                var validationRuleContainer = validationRules
-                    .FirstOrDefault(x => x.ValidationRule.ValidationRuleIdentifier == validationRuleIdentifier);
-
-                if (validationErrorTextTokens.Contains(cimValidationErrorTextToken) && validationRuleContainer != null)
-                    Assert.True(validationRuleContainer.ValidationRule is IValidationRuleWithExtendedData);
-            }
         }
 
         /// <summary>
